Alterando o tipo de unit para decimal

+++ Domain/DTOs/AssetTransactionDTO.cs
@@ -8,7 +8,7 @@ namespace TrackMyAssets_API.Domain.DTOs
     public class AssetTransactionDTO
     {
         public Guid AssetId { get; set; }
-        public double UnitsChanged { get; set; }
+        public decimal UnitsChanged { get; set; }
         public DateTime Date { get; set; }
         public string Type { get; set; } = default!;
         public string? Note { get; set; }
