+++ Domain/ModelsViews/UserAssetModelView.cs
@@ -10,6 +10,6 @@ namespace TrackMyAssets_API.Domain.ModelsViews
         // public Guid Id { get; set; }
         public Guid UserId { get; set; }
         public Guid AssetId { get; set; }
-        public double Units { get; set; }
+        public decimal Units { get; set; }
     }
 }
